+++ wordpress/wp-content/themes/codigo/resources/scripts/app.js
@@ -19,6 +19,7 @@ import accordion from "./modules/accordion";
 import jobs from "./modules/jobs";
 import salaries from "./modules/salaries";
 import loadEffect from "./modules/loadEffect";
+import magazine from "./modules/magazine";
 
 /**
  * Application entrypoint
@@ -47,8 +48,7 @@ const CDG = {
     tinyslider();
     collaborators();
     accordion();
-
-
+    magazine();
 
     window.addEventListener("resize", function(){
       //consoleHello('window has resized');