+++ wordpress/wp-content/themes/codigo/resources/views/single.blade.php
@@ -2,6 +2,10 @@
 
 @section('content')
   @while(have_posts()) @php(the_post())
-    @includeFirst(['partials.content-single-' . get_post_type(), 'partials.content-single'])
+    @if(has_block('acf/magazine'))
+      @include('partials.content-single-collaborator-magazine')
+    @else
+      @includeFirst(['partials.content-single-' . get_post_type(), 'partials.content-single'])
+    @endif
   @endwhile
 @endsection